+++ src/demo-app/app/app.routing.ts
@@ -29,6 +29,7 @@ export const routes: Routes = [
          { path: 'combobox-demo', loadChildren: '@stratio/egeo#StComboboxDemoModule' },
          { path: 'help-demo', loadChildren: '@stratio/egeo#StHelpDemoModule' },
          { path: 'info-box-demo', loadChildren: '@stratio/egeo#StInfoBoxDemoModule' },
+         { path: 'item-list-demo', loadChildren: '@stratio/egeo#StItemListDemoModule' },
          { path: 'tip-demo', loadChildren: '@stratio/egeo#StTipDemoModule' },
          { path: 'search-demo', loadChildren: '@stratio/egeo#StSearchDemoModule' }
       ]