+++ src/lib/st-item-list/st-item-list.model.ts
@@ -13,19 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export interface MenuOption {
-   route: string;
+import { TranslateableElement } from '../utils/egeo-resolver/egeo-resolve-model';
+
+export interface StItemListElement {
+   id: string | number;
    name: string;
+   icon?: string;
+   selected?: boolean;
 }
 
-export const OPTIONS: MenuOption[] = [
-   { route: 'main', name: 'main' },
-   { route: 'dropdown-menu-demo', name: 'Dropdown Menu' },
-   { route: 'dropdown-demo', name: 'Dropdown' },
-   { route: 'combobox-demo', name: 'Combobox' },
-   { route: 'help-demo', name: 'help' },
-   { route: 'info-box-demo', name: 'info-box' },
-   { route: 'item-list-demo', name: 'item-list' },
-   { route: 'tip-demo', name: 'tip' },
-   { route: 'search-demo', name: 'Search' }
-];
+export interface StItemListConfig {
+   title: string;
+   searchPlaceholder: string;
+}
+
+export interface StItemListConfigSchema {
+   title: TranslateableElement;
+   searchPlaceholder: TranslateableElement;
+}