+++ src/lib/st-item-list/st-item-list.module.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { VirtualScrollModule } from 'angular2-virtual-scroll';
+
+// Components
+import { ItemListItemComponent } from './item-list-item/item-list-item.component';
+import { ItemListScrollComponent } from './item-list-scroll/item-list-scroll.component';
+import { StItemListComponent } from './st-item-list.component';
+
+// Search module
+import { StSearchModule } from '../st-search/st-search.module';
+
+@NgModule({
+   imports: [CommonModule, StSearchModule, VirtualScrollModule],
+   declarations: [ItemListScrollComponent, ItemListItemComponent, StItemListComponent],
+   exports: [StItemListComponent]
+})
+export class StItemListModule { }